test(projects): cover trace readiness grouping
