+++ frontend/package.json
@@ -9,6 +9,7 @@
     "start": "next start",
     "lint": "eslint",
     "test": "vitest run",
+    "coverage": "vitest run src/components/project-trace-readiness.test.ts --coverage --coverage.provider=v8 --coverage.reporter=json-summary --coverage.reporter=json --coverage.include=src/components/project-trace-readiness.ts",
     "typecheck": "tsc --noEmit",
     "full:smoke": "node scripts/full-product-ui-smoke.mjs",
     "pilot:smoke": "node scripts/pilot-ui-smoke.mjs",
@@ -36,6 +37,7 @@
     "@types/node": "^26",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "@vitest/coverage-v8": "4.1.9",
     "eslint": "^9",
     "eslint-config-next": "16.2.9",
     "fast-check": "^4.8.0",
